Rename game state to task
